requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
av==14.2.0
bitsandbytes==0.45.5
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.3
cycler==0.12.1
datasets==3.5.1
debugpy==1.8.14
decorator==5.2.1
decord==0.6.0
deepspeed==0.17.5
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.13.1
fonttools==4.59.2
frozenlist==1.6.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.9
hjson==3.1.0
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
kiwisolver==1.4.9
liger_kernel==0.6.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
nest_asyncio==1.6.0
networkx==3.3
ninja==1.11.1.4
numpy==2.1.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
parso==0.8.4
peft==0.15.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pip==25.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentry-sdk==2.27.0
setproctitle==1.3.5
setuptools==79.0.1
six==1.17.0
smmap==5.0.2
stack_data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.4.0
trl==0.25.0
typing_extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
wandb==0.19.10
wcwidth==0.2.13
wheel==0.45.1
widgetsnbextension==4.0.14
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
